+++ lit_gpt/data/__init__.py
@@ -1,4 +1,6 @@
 # Copyright Lightning AI. Licensed under the Apache License 2.0, see LICENSE file.
 
-from lit_gpt.data.data_module import LitDataModule
+from lit_gpt.data.data_module import LitDataModule, LitDataModule, SFTDataset, apply_prompt_template, sft_collate_fn
+from lit_gpt.data.alpaca import Alpaca
+from lit_gpt.data.lima import LIMA
 from lit_gpt.data.tinyllama import TinyLlama
