--- lit_gpt/datasets/__init__.py
@@ -1,6 +1,0 @@
-# Copyright Lightning AI. Licensed under the Apache License 2.0, see LICENSE file.
-
-from lit_gpt.datasets.base import LitDataModule, SFTDataset, apply_prompt_template, sft_collate_fn
-from lit_gpt.datasets.alpaca import Alpaca
-from lit_gpt.datasets.lima import LIMA
-from lit_gpt.datasets.tinyllama import TinyLlama
